convert joins to global if needed, some refactoring
